rename and move frontSequelize and move ContentFragment
